refactor: 화면 별 패키지 분리
