+++ app/src/main/java/woowacourse/movie/completed/ReservationCompletedPresenter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.completed
 
 class ReservationCompletedPresenter(private val view: ReservationCompletedContract.View) :
     ReservationCompletedContract.Presenter {
