+++ app/src/main/java/woowacourse/movie/main/MainContract.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.main
 
 import woowacourse.movie.model.Movie
