+++ app/src/main/java/woowacourse/movie/reservation/ReservationContract.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.reservation
 
 import woowacourse.movie.model.Movie
