+++ app/src/main/java/woowacourse/movie/reservation/ReservationPresenter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.reservation
 
 import woowacourse.movie.model.Movie
