+++ app/src/main/res/layout/activity_reservation.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ReservationActivity">
+    tools:context=".reservation.ReservationActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
