+++ app/src/main/res/layout/activity_reservation_completed.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ReservationCompletedActivity">
+    tools:context=".completed.ReservationCompletedActivity">
 
     <TextView
         android:id="@+id/message"
